wip data generator v2

+++ schema/json/generator/v2/anyOf_test.go
@@ -20,7 +20,7 @@ func TestAnyOf(t *testing.T) {
 			},
 			test: func(t *testing.T, v interface{}, err error) {
 				require.NoError(t, err)
-				require.Equal(t, -3652171958352792229, v)
+				require.Equal(t, int64(-3652171958352792229), v)
 			},
 		},
 		{
